MOSS implemented

+++ deebo/tests_algo.py
@@ -139,12 +139,13 @@ def _test_ucb1():
 
     print("Best arm is " + str(np.argmax(means)))
 
-    algo = UCB1([], [], [])
+    algo = UCB1()
     algo.reset(n_arms)
     results = test_algorithm(algo, arms, 1000, 250)
-    filename = 'ucb1.csv'
+    filename = 'ucb1_test.csv'
     fp = './logs/scenario2/optim/' + filename
-    results.to_csv(fp)
+    if input('save result (might overwrite)? : ') == 'y':
+        results.to_csv(fp)
 
     return
 
@@ -167,6 +168,31 @@ def _test_ucb1_tuned():
     return
 
 
+def _test_moss(scenario=1):
+
+    if scenario == 1:
+        means = [0.1, 0.2, 0.3, 0.4, 0.9]
+    elif scenario == 2:
+        means = [0.1, 0.1, 0.1, 0.1, 0.2]
+    elif scenario == 3:
+        means = [0.1, 0.25, 0.5, 0.75, 0.9]
+    else:
+        exit(1)
+
+    n_arms = len(means)
+    arms = list(map(lambda x: BernoulliArm(x), means))
+
+    print("Best arm is " + str(np.argmax(means)))
+
+    algo = MOSS()
+    algo.reset(n_arms)
+    results = test_algorithm(algo, arms, 1, 250)
+    filename = f'moss.csv'
+    fp = f'./logs/scenario{scenario}/optim/' + filename
+    if input('save result (might overwrite)? : ') == 'y':
+        results.to_csv(fp)
+
+
 def _test_etc(scenario=1):
 
     if scenario == 1:
@@ -183,7 +209,7 @@ def _test_etc(scenario=1):
 
     print("Best arm is " + str(np.argmax(means)))
 
-    exp_len = [2,3,4,6,7,8,9]
+    exp_len = np.arange(15)+1
 
     for e in exp_len:
         algo = ETC([], [], e)
@@ -274,4 +300,4 @@ def _test_exp3(scenario=3):
 
 
 if __name__ == '__main__':
-    _test_etc()
+    _test_moss()
